Update the effects of changing quantity of stocks to double.
Updated the ExcelBuilder class.

+++ src/main/java/com/boheco1/dev/integratedaccountingsystem/dao/ReleasingDAO.java
@@ -18,7 +18,7 @@ public class ReleasingDAO {
 
         ps.setString(1, releasing.getStockID());
         ps.setString(2, releasing.getMirsID());
-        ps.setInt(3, releasing.getQuantity());
+        ps.setDouble(3, releasing.getQuantity());
         ps.setDouble(4, releasing.getPrice());
         ps.setString(5, releasing.getUserID());
         ps.setString(6, releasing.getStatus());
